backend/requirements.txt
# Core
fastapi
uvicorn[standard]
motor
pydantic-settings
python-dotenv
pymongo

# Testing
pytest
pytest-asyncio
httpx

# Database
pandas

scikit-learn
requests
